fix!: selecting multiple files only opens buffers, not windows (#111)
In an oil buffer you can select multiple entries by highlighting them in
visual mode and then calling `oil.select()` (or using a keymap that does
so, like `<CR>` or `<C-s>`. Previously we would automatically assume
that you wanted to open multiple entries as splits. Now, we treat a
default call to `select()` (which is what `<CR>` does) as if you simply
wanted to open the buffers and keep the same window. This plays nicer
with bufferlines. If you want the old behavior, you can explicitly pass
in `oil.select({vertical = true})` or use the `<C-s>` keymap when you
have multiple entries selected.