refactor: set concealcursor=nvic by default
Originally I set the default value to `n` because I wanted users to be
aware of the magic ID number at the start of the line if they were
editing a buffer. With the recent improvements to the cursor placement,
I think that it's no longer necessary. Most normal edits should just
work as expected now, and it's a nicer experience if the visuals don't
jump around when you enter visual/insert mode.